+++ demo/index.js
@@ -1,15 +1,24 @@
-var CODEMIRROR_ROOT = "../node_modules/codemirror/";
-var HYPERMD_ROOT = "../hypermd/";
+if (requirejs) requirejs.config({
+    paths: {
+        "codemirror": "./node_modules/codemirror/",
+        "hypermd": "./hypermd/"
+    },
+    waitSeconds: 15
+})
 
-var DEBUG_PATCHING = false  // patching: show/hide tokens
+// Using requirejs's "path" option (see above).
+// Change these if things are different
+var CODEMIRROR_ROOT = window.CODEMIRROR_ROOT = "codemirror/";
+var HYPERMD_ROOT = window.HYPERMD_ROOT = "hypermd/";
 
-requirejs([
+require([
     CODEMIRROR_ROOT + 'lib/codemirror',
     CODEMIRROR_ROOT + 'mode/javascript/javascript',
     CODEMIRROR_ROOT + 'addon/fold/foldcode',
     CODEMIRROR_ROOT + 'addon/fold/foldgutter',
     CODEMIRROR_ROOT + 'addon/fold/markdown-fold',
     CODEMIRROR_ROOT + 'addon/edit/continuelist',
+    // CODEMIRROR_ROOT + 'keymap/vim',
     HYPERMD_ROOT + 'mode/hypermd',
     HYPERMD_ROOT + 'addon/hide-token',
     HYPERMD_ROOT + 'addon/cursor-debounce',
@@ -19,31 +28,25 @@ requirejs([
     HYPERMD_ROOT + 'addon/hover'
 ], function (CodeMirror) {
     'use strict';
-    var myTextarea = document.getElementById('demo');
+    var myTextarea = document.getElementById('demo')
 
-    var xmlhttp, url = "README.md";
-    if (window.XMLHttpRequest) { xmlhttp = new XMLHttpRequest() }
-    else if (window.ActiveXObject) { xmlhttp = new ActiveXObject("Microsoft.XMLHTTP") }
-    if (xmlhttp != null) {
-        xmlhttp.onreadystatechange = function () {
-            if (xmlhttp.readyState == 4 && xmlhttp.status == 200) {
-                myTextarea.innerText = xmlhttp.responseText
-                editor_init()
-            }
-        }
-        xmlhttp.open("GET", url, true)
-        xmlhttp.send(null)
-    }
+    //init_editor()
+    ajax_load_file_then_init_editor("README.md")
 
-    function editor_init() {
+    function init_editor() {
         var editor = CodeMirror.fromTextArea(myTextarea, {
             lineNumbers: true,
             lineWrapping: true,
             theme: "hypermd-light",
             mode: "text/x-hypermd",
+            // keyMap: "vim",
 
             foldGutter: true,
-            gutters: ["CodeMirror-linenumbers", "CodeMirror-foldgutter"],
+            gutters: [
+                "CodeMirror-linenumbers",
+                "CodeMirror-foldgutter",
+                "HyperMD-goback"
+            ],
             extraKeys: {
                 "Enter": "newlineAndIndentContinueMarkdownList"
             },
@@ -53,11 +56,26 @@ requirejs([
         });
 
         editor.hmdHideTokenInit()
-        editor.hmdClickInit()
         editor.hmdHoverInit()
 
         window.CodeMirror = CodeMirror
         window.editor = editor
         editor.setSize("100%", 500)
+    }
+
+    function ajax_load_file_then_init_editor(url) {
+        var xmlhttp;
+        if (window.XMLHttpRequest) { xmlhttp = new XMLHttpRequest() }
+        else if (window.ActiveXObject) { xmlhttp = new ActiveXObject("Microsoft.XMLHTTP") }
+        if (xmlhttp != null) {
+            xmlhttp.onreadystatechange = function () {
+                if (xmlhttp.readyState == 4 && xmlhttp.status == 200) {
+                    myTextarea.innerText = xmlhttp.responseText
+                    init_editor()
+                }
+            }
+            xmlhttp.open("GET", url, true)
+            xmlhttp.send(null)
+        }
     }
 })